Cleaned recipient.py of static methods and moved it to types package, adjusted recipient_gateway imports accordingly and added more inline comments.
