+++ trolley/__init__.py
@@ -1,7 +1,7 @@
 from trolley.configuration import Configuration
 from trolley.client import Client
 from trolley.gateway import Gateway
-from trolley.recipient import Recipient
+from trolley.types.recipient import Recipient
 from trolley.types.recipient_account import RecipientAccount
 from trolley.types.log import Log
 from trolley.types.batch import Batch
