+++ trolley/recipient_gateway.py
@@ -1,7 +1,10 @@
 from collections import namedtuple
 from trolley.exceptions.invalidFieldException import InvalidFieldException
 import trolley.configuration
+from trolley.types.log import Log
 from trolley.types.meta import Meta
+from trolley.types.payment import Payment
+from trolley.types.recipient import Recipient
 from trolley.utils.url_utils import UrlUtils
 
 
@@ -15,13 +18,18 @@ class RecipientGateway(object):
         self.gateway = gateway
         self.config = config
 
+    """
+        Retrieve a recipient
+            A recipient_id is required::
+            recipient.find('R-fjeracjmuflh')
+        """
     def find(self, recipient_id, term=""):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
         endpoint = f'/v1/recipients/{recipient_id}/{term}'
         response = trolley.configuration.Configuration.client(
             self.config).get(endpoint)
-        recip = trolley.Recipient.factory(response)
+        recip = Recipient.factory(response)
         recipient = namedtuple("Recipient", recip.keys())(*recip.values())
         count = 0
         for account in recipient.accounts:
@@ -30,16 +38,27 @@ class RecipientGateway(object):
             count = count + 1
         return recipient
 
+    """
+        Create a recipient
+            A body is required::
+            recipient.create({"type": "individual", "firstName": "John",
+                                    "lastName": "Smith", "email": "jh@edxample.com"})
+        """
     def create(self, body):
         if body is None:
             raise InvalidFieldException("Body cannot be None.")
         endpoint = '/v1/recipients/'
         response = trolley.configuration.Configuration.client(
             self.config).post(endpoint, body)
-        recip = trolley.Recipient.factory(response)
+        recip = Recipient.factory(response)
         recipient = namedtuple("Recipient", recip.keys())(*recip.values())
         return recipient
 
+    """
+        Update a recipient
+            A recipient_id and body are required::
+            recipient.update('R-fjeracjmuflh', {'firstName': 'tom'})
+        """
     def update(self, recipient_id, body):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
@@ -50,6 +69,11 @@ class RecipientGateway(object):
             self.config).patch(endpoint, body)
         return True
 
+    """
+        Delete a recipient
+            A recipient_id is required::
+            recipient.delete('R-fjeracjmuflh')
+        """
     def delete(self, recipient_id):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
@@ -58,6 +82,14 @@ class RecipientGateway(object):
             self.config).delete(endpoint)
         return True
     
+    """
+        Delete multiple recipients
+            A list of recipient_ids is required::
+            recipient.delete_multiple([
+                'R-fjeracjmuflh',
+                'R-ajerasdfuflf'
+            ])
+        """
     def delete_multiple(self, recipient_ids):
         if recipient_ids is None:
             raise InvalidFieldException("Recipient ids cannot be None.")
@@ -67,6 +99,11 @@ class RecipientGateway(object):
             self.config).delete(endpoint, recipient_ids)
         return True
 
+    """
+        Retrieve Recipient activity logs, by page
+            A recipient_id is required:
+            recipient.retrieve_logs('R-fjeracjmuflh')
+        """
     def retrieve_logs(self, recipient_id, page=1, pageSize=10):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
@@ -78,7 +115,7 @@ class RecipientGateway(object):
         logs = []
         count = 0
         for log in response['recipientLogs']:
-            temp = trolley.Log.factory(log)
+            temp = Log.factory(log)
 
             logs.insert(count, namedtuple("Log", temp.keys())(*temp.values()))
 
@@ -89,6 +126,11 @@ class RecipientGateway(object):
 
         return logs
     
+    """
+        Retrieve all payments of the Recipient
+            A recipient_id is required:
+            recipient.get_all_payments('R-fjeracjmuflh')
+        """
     def get_all_payments(self, recipient_id):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
@@ -99,7 +141,7 @@ class RecipientGateway(object):
         payments = []
         count = 0
         for payment in response['payments']:
-            temppayment = trolley.Payment.factory(payment)
+            temppayment = Payment.factory(payment)
             newpayment = namedtuple("Payment", temppayment.keys())(*temppayment.values())
             payments.insert(count, newpayment)
             count = count + 1
@@ -109,6 +151,11 @@ class RecipientGateway(object):
 
         return payments
 
+    """
+        Retrieve all offline payments of the Recipient
+            A recipient_id is required:
+            recipient.get_all_offline_payments('R-fjeracjmuflh')
+        """
     def get_all_offline_payments(self, recipient_id):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
@@ -119,7 +166,7 @@ class RecipientGateway(object):
         offlinePayments = []
         count = 0
         for offlinePayment in response['offlinePayments']:
-            temppayment = trolley.Payment.factory(offlinePayment)
+            temppayment = Payment.factory(offlinePayment)
             newpayment = namedtuple("Payment", temppayment.keys())(*temppayment.values())
             offlinePayments.insert(count, newpayment)
             count = count + 1
@@ -134,7 +181,8 @@ class RecipientGateway(object):
         You can use this generator in a foreach loop to sequentially go through all the 
         search results without needing to call this method again.
         
-            For accessing specific pages, check the search_by_page() method. """
+        For accessing specific pages, check the search_by_page() method. 
+        """
     def search(self, search=None, name=None, email=None, reference_id=None, start_date=None,
                end_date=None, status=None, compliance_status=None, country=None, payout_method=None, currency=None,
                order_by=None, sort_by=None):
@@ -156,7 +204,9 @@ class RecipientGateway(object):
     
     """ Search Recipients by providing a page number.
         This method returns a list.
-        You should use this function when you want to paginate manually. """
+        You should use this function when you want to paginate manually.
+        recipient.search_by_page(1,10,'test')
+     """
     def search_by_page(self, page=1, page_size=10, search=None, name=None, email=None, reference_id=None, start_date=None,
                end_date=None, status=None, compliance_status=None, country=None, payout_method=None, currency=None,
                order_by=None, sort_by=None):
@@ -169,7 +219,7 @@ class RecipientGateway(object):
         recipients = []
         count = 0
         for recipient in response['recipients']:
-            temp = trolley.recipient.Recipient.factory(recipient)
+            temp = Recipient.factory(recipient)
 
             recipient = namedtuple("Recipient", temp.keys())(*temp.values())
             recipients.insert(count, recipient)
